# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FixUp identity API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/requests" : {
      "post" : {
        "tags" : [ "repair-request-controller" ],
        "summary" : "Open a repair request",
        "description" : "The body carries media asset IDs. No raw image binary content crosses this API.",
        "operationId" : "open",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OpenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid specialty, missing field or unexpected client fields",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid bearer token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Inactive account, missing role or request not visible to the caller",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The repair request does not exist",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The transition does not apply to the current state of the request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201" : {
            "description" : "The request is open and visible to the fixers",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestDetailResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/requests/{requestId}" : {
      "get" : {
        "tags" : [ "repair-request-controller" ],
        "summary" : "Read one repair request",
        "description" : "The owner always reads it; a verified compatible fixer reads it while open, or afterwards only when the work was assigned to him.",
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid specialty, missing field or unexpected client fields",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid bearer token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Inactive account, missing role or request not visible to the caller",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The repair request does not exist",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The transition does not apply to the current state of the request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "The request with its description and photos",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestDetailResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/requests/open" : {
      "get" : {
        "tags" : [ "repair-request-controller" ],
        "summary" : "List the open requests offered to fixers",
        "description" : "FR-UC-18: the fixer's inbox. Requires active, verified FIXER and filters by fixer specialties.",
        "operationId" : "open_1",
        "responses" : {
          "400" : {
            "description" : "Invalid specialty, missing field or unexpected client fields",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid bearer token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Inactive account, missing role or request not visible to the caller",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The repair request does not exist",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The transition does not apply to the current state of the request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Open requests matching fixer specialties, newest first",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OpenRequestSummaryResponse"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/requests/me" : {
      "get" : {
        "tags" : [ "repair-request-controller" ],
        "summary" : "List the repair requests opened by the current user",
        "operationId" : "mine",
        "responses" : {
          "400" : {
            "description" : "Invalid specialty, missing field or unexpected client fields",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid bearer token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Inactive account, missing role or request not visible to the caller",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "The repair request does not exist",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The transition does not apply to the current state of the request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Requests owned by the current user, newest first",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RequestDetailResponse"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "OpenRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "specialty" : {
            "$ref" : "#/components/schemas/Specialty"
          },
          "title" : {
            "type" : "string",
            "maxLength" : 150,
            "minLength" : 0
          },
          "description" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "mediaIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "maxItems" : 6,
            "minItems" : 0
          }
        },
        "required" : [ "description", "specialty", "title" ]
      },
      "Specialty" : {
        "type" : "string",
        "enum" : [ "PLUMBING", "ELECTRICAL", "PAINTING", "CARPENTRY", "MASONRY", "GENERAL" ]
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          }
        }
      },
      "PhotoResponse" : {
        "type" : "object",
        "properties" : {
          "mediaId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "readUrl" : {
            "type" : "string"
          },
          "readUrlExpiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "mediaId", "readUrl", "readUrlExpiresAt" ]
      },
      "RepairRequestStatus" : {
        "type" : "string",
        "enum" : [ "OPEN", "ASSIGNED" ]
      },
      "RequestDetailResponse" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "specialty" : {
            "$ref" : "#/components/schemas/Specialty"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "photos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PhotoResponse"
            }
          },
          "status" : {
            "$ref" : "#/components/schemas/RepairRequestStatus"
          },
          "assignedFixerUserId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "description", "photos", "requestId", "specialty", "status", "title" ]
      },
...
      "OpenRequestSummaryResponse" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "specialty" : {
            "$ref" : "#/components/schemas/Specialty"
          },
          "title" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "requestId", "specialty", "title" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}